requirements.txt
python-dotenv~=1.0.1
Flask~=3.0.3
pymongo~=4.7.2
Werkzeug~=3.0.3